Circle Parameters
Write a Java program that generates and prints circle-related values. Read in a value for radius. Calculate and print the corresponding values for diameter, circumference, and area. Use the built-in Java constant Math.PI for the value of π. Use Math.PI the same way that you would use 3.1452 in your program –there’s no need to declare Math.PI with a declaration statement, just use it.

Sample session:

Enter a radius value: 2.5
Diameter = 5.0
Circumference = 15.707963267948966
Area = 19.634954084936208